+++ ch-user-api/src/main/java/com/ai/ch/user/api/shopinfo/params/InsertShopInfoRequst.java
@@ -5,8 +5,7 @@ import java.sql.Timestamp;
 import com.ai.opt.base.vo.BaseInfo;
 
 /**
- * 保存店铺信息服务
- * Date: 2016年8月23日 <br>
+ * 保存店铺信息服务 Date: 2016年8月23日 <br>
  * Copyright (c) 2016 asiainfo.com <br>
  * 
  * @author zhangqiang7
@@ -43,7 +42,7 @@ public class InsertShopInfoRequst extends BaseInfo {
 	/**
 	 * 订单划扣比例
 	 */
-	private Integer ratio;
+	private Float ratio;
 
 	/**
 	 * 店铺等级
@@ -136,7 +135,7 @@ public class InsertShopInfoRequst extends BaseInfo {
 	public void setMerchantNo(String merchantNo) {
 		this.merchantNo = merchantNo;
 	}
-	
+
 	public String getUserId() {
 		return userId;
 	}
@@ -177,11 +176,11 @@ public class InsertShopInfoRequst extends BaseInfo {
 		this.rentFee = rentFee;
 	}
 
-	public Integer getRatio() {
+	public Float getRatio() {
 		return ratio;
 	}
 
-	public void setRatio(Integer ratio) {
+	public void setRatio(Float ratio) {
 		this.ratio = ratio;
 	}
